Identity (#4)
* Prepared implementation of identity framework

* Resolved issue with package

* Scaffolded identity

* Added pages to login and register

* Added authorize attribute to caps conroller

* Index only displays caps for current user

* Added method to read cap based on the current user

* Added TarczynewsUser repository

* Injected user repository and changed actions to work on user caps rather than on all caps

* Fixed deletion

* Fixed bug with editing

+++ Controllers/CapsController.cs
@@ -1,31 +1,32 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Tarczynews.Models;
-using System.Linq;
-using Tarczynews.Data;
 using Tarczynews.Repositories;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Tarczynews.Controllers
 {
+    [Authorize]
     public class CapsController : Controller
     {
         private readonly ITarczynCapRepository _tarczynCapRepository;
+        private readonly ITarczynewsUserRepository _tarczynewsUserRepository;
 
-        public CapsController(ITarczynCapRepository tarczynCapRepository)
+        public CapsController(ITarczynCapRepository tarczynCapRepository, ITarczynewsUserRepository tarczynewsUserRepository)
         {
             _tarczynCapRepository = tarczynCapRepository;
+            _tarczynewsUserRepository = tarczynewsUserRepository;
         }
 
         // GET: CapsController
         public ActionResult Index()
         {
-            return View(_tarczynCapRepository.ReadAllTarczynCapsSortedAscendingByNumber());
+            return View(_tarczynCapRepository.ReadAllTarczynCapsForUsernameSortedAscendingByNumber(_tarczynewsUserRepository.ReadCurrent().UserName));
         }
 
         // GET: CapsController/Details/5
         public ActionResult Details(int number)
         {
-            var cap = _tarczynCapRepository.ReadTarczynCapByNumber(number);
+            var cap = _tarczynCapRepository.ReadTarczynCapByNumberAndOwnerUsername(number, _tarczynewsUserRepository.ReadCurrent().UserName);
 
             if (cap == null)
             {
@@ -48,7 +49,14 @@ namespace Tarczynews.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create(TarczynCap tarczynCap)
         {
-            var storedCap = _tarczynCapRepository.ReadTarczynCapByNumber(tarczynCap.Number);
+            var user = this._tarczynewsUserRepository.ReadCurrent();
+            if (user == null || string.IsNullOrEmpty(user.UserName))
+            {
+                TempData["Error"] = "Unauthorized access. Try to log in again";
+
+                return RedirectToAction("Index", "Home");
+            }
+            var storedCap = _tarczynCapRepository.ReadTarczynCapByNumberAndOwnerUsername(tarczynCap.Number, user.UserName);
 
             if(storedCap != null)
             {
@@ -57,8 +65,21 @@ namespace Tarczynews.Controllers
                 return View(tarczynCap);
             }
 
+            tarczynCap.Owner = user;
+            
+            if(user.TarczynCaps != null)
+            {
+                user.TarczynCaps.Add(tarczynCap);
+            }
+            else
+            {
+                user.TarczynCaps = new List<TarczynCap>() { tarczynCap };
+            }
+
             _tarczynCapRepository.Create(tarczynCap);
             _tarczynCapRepository.Save();
+
+            _tarczynewsUserRepository.Update(user);
 
             TempData["Success"] = $"Cap {tarczynCap.Number} was created successfully";
 
@@ -68,7 +89,7 @@ namespace Tarczynews.Controllers
         // GET: CapsController/Edit/5
         public ActionResult Edit(int number)
         {
-            var cap = _tarczynCapRepository.ReadTarczynCapByNumber(number);
+            var cap = _tarczynCapRepository.ReadTarczynCapByNumberAndOwnerUsername(number, _tarczynewsUserRepository.ReadCurrent().UserName);
 
             if (cap == null)
             {
@@ -86,7 +107,7 @@ namespace Tarczynews.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult EditPost(TarczynCap tarczynCap)
         {
-            var storedCap = _tarczynCapRepository.Read(tarczynCap.Id);
+            var storedCap = _tarczynCapRepository.ReadTarczynCapByIdAndOwnerUsername(tarczynCap.Id, _tarczynewsUserRepository.ReadCurrent().UserName);
 
             if(storedCap == null)
             {
@@ -96,7 +117,7 @@ namespace Tarczynews.Controllers
             {
                 if (storedCap.Number != tarczynCap.Number)
                 {
-                    var storedCapWithNumber = _tarczynCapRepository.ReadTarczynCapByNumber(tarczynCap.Number);
+                    var storedCapWithNumber = _tarczynCapRepository.ReadTarczynCapByNumberAndOwnerUsername(tarczynCap.Number, _tarczynewsUserRepository.ReadCurrent().UserName);
 
                     if (storedCapWithNumber != null)
                     {
@@ -118,7 +139,7 @@ namespace Tarczynews.Controllers
         // GET: CapsController/Delete/5
         public ActionResult Delete(int number)
         {
-            var cap = _tarczynCapRepository.ReadTarczynCapByNumber(number);
+            var cap = _tarczynCapRepository.ReadTarczynCapByNumberAndOwnerUsername(number, _tarczynewsUserRepository.ReadCurrent().UserName);
 
             if (cap == null)
             {
@@ -137,7 +158,7 @@ namespace Tarczynews.Controllers
         public ActionResult DeletePost(Guid id)
         {
             var cap = _tarczynCapRepository.Read(id);
-            if (cap != null)
+            if (cap != null && _tarczynewsUserRepository.ReadCurrent().Id.Equals(cap.Owner?.Id ?? ""))
             {
                 var number = cap.Number;
                 _tarczynCapRepository.Delete(id);
